requirements.txt
aiohttp==3.8.4
aiosqlite==0.19.0
discord.py==2.2.3
parsedatetime==2.6
psutil==5.9.4
python_dateutil==2.8.2
Requests==2.31.0
typing_extensions==4.6.3
Wavelink==2.3.1
yarl==1.8.2